# perf!: faster image preview with optimized `magick` arguments