@sxyazi 2025-03-25
That's incredible! 

Can you explain to me the reason for each parameter change?