@ze0987 2025-03-25
The order of some options is crucial, for example, `-flatten` at the beginning is slower than at the end (when working on a scaled-down image). `-density` unnecessarily increases the size of the thumbnail. We want `-auto-orient` before deleting metadata with `-strip`. `-sample` is much faster than `-resize`.