@sxyazi 2025-03-25
Moving `-flatten` to the end will break multi-level images. 

You can download `test.zip` from https://github.com/sxyazi/yazi/pull/1684#issuecomment-2375081530, and then try:

```
magick test.xcf -auto-orient -strip -sample 500x -flatten -quality 90 JPG:test.jpg
```

and

```
magick test.xcf -auto-orient -strip -flatten -sample 500x -quality 90 JPG:test.jpg
```