@ze0987 2025-03-25
I actually checked this during testing. I didn't notice that one layer was missing. Sorry about that.

`-flatten` before scaling is quite expensive. Can we use it conditionally? For example, for selected mime types or with option in `yazi.toml` disabled by default?

```
Benchmark 1: taskset -c 19 magick IMG_001.jxl -auto-orient -strip -sample 600x -quality 75 -flatten JPG:IMG_001_th
  Time (mean ± σ):      1.705 s ±  0.002 s    [User: 1.629 s, System: 0.060 s]
  Range (min … max):    1.702 s …  1.708 s    10 runs
 ```
```
18:57 zin ~/Downloads/JXL $  hyperfine -w 1 -r 10 'taskset -c 19 magick IMG_001.jxl -auto-orient -strip -flatten -sample 600x -quality 75 JPG:IMG_001_th'
Benchmark 1: taskset -c 19 magick IMG_001.jxl -auto-orient -strip -flatten -sample 600x -quality 75 JPG:IMG_001_th
  Time (mean ± σ):      2.881 s ±  0.003 s    [User: 2.762 s, System: 0.094 s]
  Range (min … max):    2.876 s …  2.886 s    10 runs
```