@sxyazi 2025-03-25
Maybe add a `--flatten` parameter to enable it manually when the user needs it, such as:

```
{ name = "*.xcf", run = "magick --flatten" }
```

Then receive it with `job.args.flatten` in the `preload()` method.

So that we don't have to incur ongoing performance loss for all image formats, since images with multi-layer didn't seem very common