@ze0987 2025-03-25
Awesome. Unfortunately, I will need assistance with this, or some time to figure out how to implement it, since learning Lua is still only in my plans.

By the way, I noticed that rust previewer does not create thumbnails on hover when the preloader is disabled. Is this an intended behavior?