@sxyazi 2025-03-26
> I will need assistance with this

No worries, I'll take care of it

> I noticed that rust previewer

What's rust previewer? You mean `magick` previewer or what?