@ze0987 2025-03-26
> No worries, I'll take care of it

Awesome. Thank you!

> What's rust previewer? You mean `magick` previewer or what?

`image` if I'm not mistaken (?) It handles jpeg, webp, among others. With image preloaders disabled, please clear the cache, hover over any jpg/png/webp file, check the cache for the thumbnail.

```
prepend_preloaders = [
    { mime = "image/*", run = "noop" },
}
```