@ze0987 2025-03-26
I was wrong again. `-density` is not totally useless at all because it is negated by `-sample`. It can actually "break" things. Source svg: https://pl.wikipedia.org/wiki/Plik:NewTux.svg

w/o `-density`:
![wodensity](https://github.com/user-attachments/assets/09cba1fa-ea34-46fb-95da-7cb4d7cc2170)

with `-density`:
![density](https://github.com/user-attachments/assets/0a219948-dc18-44f1-9cb7-82cc399a8a26)