@sxyazi 2025-03-26
> `image` if I'm not mistaken (?) It handles jpeg, webp, among others. With image preloaders disabled, please clear the cache, hover over any jpg/png/webp file, check the cache for the thumbnail.

Yes, this is expected, `image` previewer uses the builtin image decoder, so there is no need for extra format conversion like that in `magick`, which would require creating an intermediate file. If the user has not enabled caching, the preview is displayed directly from the source file.

> I was wrong again. `-density` is not totally useless at all because it is negated by `-sample`. It can actually "break" things. Source svg: 

Should be fixed now in the latest commit – changed back to `-resize` from `-sample` for SVG.