@sxyazi 2025-03-26
In my tests, not adding `-density` causes the resulting image to be blurry. 

Original SVG:

![SVG_example_markup_grid](https://github.com/user-attachments/assets/78e23c13-0c3f-45b7-a5dc-45d31e472aa2)

## With `-density`:

```sh
magick -density 200 SVG_example_markup_grid.svg -strip -resize 1000x1000^ -quality 90 density.jpg
```

![density](https://github.com/user-attachments/assets/bd978297-cdaf-4840-a4ec-65fcba7392cc)

## Without `-density`:

```sh
magick SVG_example_markup_grid.svg -strip -resize 1000x1000^ -quality 80 no-density.jpg
```

![no-density](https://github.com/user-attachments/assets/c0a3b5b6-41cc-4399-a2bc-ec23b19f3d09)