@sxyazi 2025-03-26
Oops, I passed two different `-quality` arguments - will re-test it using the same `-quality` value.