@sxyazi 2025-03-26
Edited https://github.com/sxyazi/yazi/pull/2533#issuecomment-2754416399 with the same `-quality 90`, but the result didn't change much - without `-density`, the image still appears blurry.