@ze0987 2025-03-26
Because you are up-scaling the bitmap with `^`. Please consider [`-size`](https://github.com/sxyazi/yazi/pull/2533#issuecomment-2754385976) with `>` before the svg file. Its much faster anyway.