@sxyazi 2025-03-26
That's intentional, SVG is a vector format, which means it doesn't have a fixed width or height, and its initial canvas size can be very small, which makes it less ideal for previewing.