@JustForFun88 2025-03-26
@sxyazi Maybe it would be better to switch to [`resvg`](https://github.com/linebender/resvg) for rendering SVGs? It’s similar to the `image` crate, but for the SVG format. It can render many more SVG images correctly than `magick`. [Typst](https://typst.app/) uses it as the default.

Also, rendering with resvg would be significantly faster.
If you're interested, I could prepare a PR sometime in the coming weeks (maybe a month or a bit more — just a bit busy with work right now 😅).