@sxyazi 2025-03-26
> I'll test `-size`

No luck, `-size` resizes the canvas not the actual vector. Updated https://github.com/sxyazi/yazi/pull/2533#issuecomment-2754416399