@sxyazi 2025-03-26
> Maybe it would be better to switch to [resvg](https://github.com/linebender/resvg) for rendering SVGs?

I haven't looked into `resvg` yet, not sure about the impact on binary size and build speed. It seems to have binary packages and perhaps can be invoked as a process like `magick`, but not sure if mainstream distributions have it, especially on Windows and Android.