@ze0987 2025-03-26
> That's intentional,

I understand.  I thought we desired the original size when the source is smaller than the thumbnail. In this case just the `-size` `"%dx%d^"` before `job.file.url` should do.

> No luck, -size resizes the canvas not the actual vector. Updated #2533 (comment)

Strange, here it looks like this. `magick -size '1000x1000>' image.svg -strip -quality 90 size.jpg`

![size](https://github.com/user-attachments/assets/47a58d03-c784-4cc7-bd2e-6cf4fbe9f295)

```
magick --version
Version: ImageMagick 7.1.1-46 Q16-HDRI x86_64 22747 https://imagemagick.org
Copyright: (C) 1999 ImageMagick Studio LLC
License: https://imagemagick.org/script/license.php
Features: Cipher DPC HDRI Modules OpenCL OpenMP(4.5) 
Delegates (built-in): bzlib cairo djvu fftw fontconfig freetype heic jbig jng jp2 jpeg jxl lcms lqr ltdl lzma openexr pangocairo png raqm raw rsvg tiff uhdr webp wmf x xml zip zlib zstd
Compiler: gcc (14.2)
```